improve documentation and tests to demonstrate reflect_from on a row object
the reflect_from() method can take a DBIx::Class row object instead of a
resultsource. This patch adds a touch of documentation to make that obvious.
It also adds some tests to the test suite to enforce it.

There is an optional dependency added on Test::Differences. If Test::Differences
is present, the output from comparing the HTML is a little easier to follow
if it fails.